requirements.txt
absl-py==0.12.0
alembic==0.9.9
appdirs==1.4.4
astunparse==1.6.3
autopep8==1.5.4
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.2
cffi==1.14.5
chardet==3.0.4
click==6.7
cycler==0.10.0
deepface==0.0.51
distlib==0.3.1
fer==20.1.3
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flatbuffers==1.12
gast==0.3.3
gdown==3.13.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.11.3
idna==2.6
importlib-metadata==3.3.0
itsdangerous==0.24
Jinja2==2.10.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy==1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib
mtcnn==0.1.0
numpy ==1.19.5
oauth2==1.9.0.post1
oauthlib==2.0.7
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2021.1
PyYAML==5.4.1
requests==2.21.0
requests-oauthlib==0.8.0
retina-face==0.0.4
rsa==4.7.2
scipy
six==1.15.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.22
URLObject==2.4.3
virtualenv==20.2.2
Werkzeug==0.15.0
wrapt==1.12.1
WTForms==2.1
zipp==3